seg head refactor into a new file
