+++ keras_cv/models/segmentation/__init__.py
@@ -12,5 +12,4 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-from keras_cv.models.segmentation.__internal__ import SegmentationHead
 from keras_cv.models.segmentation.deeplab import DeepLabV3
